+++ cryptol-remote-api/src/CryptolServer/Exceptions.hs
@@ -10,6 +10,7 @@ module CryptolServer.Exceptions
   , proverError
   , cryptolParseErr
   , cryptolError
+  , tcSolverTimeout
   , moduleNotLoaded
   , configLoadError
   , noModule
@@ -21,7 +22,12 @@ import qualified Data.Vector as Vector
 
 import Cryptol.Parser.AST(ModName)
 import Cryptol.ModuleSystem (ModuleError(..), ModuleWarning(..))
+import Cryptol.ModuleSystem.Monad (ImportSource(FromModule))
 import Cryptol.ModuleSystem.Name as CM
+import Cryptol.Parser.Position (emptyRange)
+import Cryptol.TypeCheck.Error (Error(TCSolverTimeout))
+import Cryptol.TypeCheck.PP (emptyNameMap)
+import Cryptol.Utils.Ident (interactiveName)
 import Cryptol.Utils.PP (pretty, PP)
 
 import Data.Aeson hiding (Encoding, Value, decode)
@@ -211,6 +217,16 @@ cryptolParseErr expr err =
   makeJSONRPCException
     20000 "Cryptol parse error"
     (Just $ JSON.object ["input" .= expr, "error" .= show err])
+
+tcSolverTimeout :: Int -> JSONRPCException
+tcSolverTimeout seconds =
+  -- Use the same error representation as a timeout caught during type-checking.
+  cryptolError
+    (TypeCheckingFailed
+      (FromModule interactiveName)
+      emptyNameMap
+      [(emptyRange, TCSolverTimeout seconds)])
+    []
 
 -- The standard way of presenting a type: a structured type, plus a
 -- human-readable string